With this type conversion on the MIK32 AMUR board, it does not work, …

+++ AGS10.cpp
@@ -33,7 +33,7 @@ void AGS10::begin()
 void AGS10::calibrateFact()
 {
   uint8_t data[4] = {CALIB_CMD1, CALIB_CMD2, CALIB_FACT, CALIB_FACT};
-  _crc = Calc_CRC8(uint8_t(data), uint8_t(0x01));
+  _crc = Calc_CRC8(data, 0x01);
   Wire.beginTransmission(AGS10_ADDR);
   Wire.write(CALIB_REG);
   Wire.write(CALIB_CMD1);
